Moved the Kernel.php file into the Infra namespace
